pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "csfloat-market-harness"
version = "0.1.0"
description = "Test harness y CLI tipado para CSFloat Market API (listings, detalle y publicación)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Pentagoo", email = "dev@pentagoo.example" }]
license = { text = "MIT" }
dependencies = [
  "httpx>=0.25,<0.28",
  "pydantic>=2.2,<3",
  "typer>=0.12,<1.0",
  "rich>=13.0,<14.0",
  "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
test = [
  "pytest>=7,<9",
  "pytest-cov>=4,<6",
  "respx>=0.20,<0.22",
]

[project.urls]
Homepage = "https://csfloat.com"
Documentation = "https://docs.csfloat.com/#introduction"

[project.scripts]
csf = "csfloat_client.cli:main"

[tool.pytest.ini_options]
addopts = "-q --cov=csfloat_client --cov-report=term-missing --cov-fail-under=70"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["csfloat_client"]

[tool.coverage.report]
show_missing = true

[tool.hatch.build.targets.wheel]
packages = ["csfloat_client"]

[tool.hatch.build.targets.sdist]
include = [
  "csfloat_client/",
  "examples/",
  "tests/",
  "README.md",
  ".env.example",
  "pyproject.toml",
]
